queen_app/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.1
python-telegram-bot==20.7
sqlite-utils==3.36
